feat: Blocking drive cache seems to work
I'm no longer looking ahead to see if the current node will overflow the RAM budget. Instead I'm just looking if the budget has been overflown before calculating the next image, and if so forcing some images into drive storage.

The reason is that it's hard to know for sure how much RAM a claculated node is going to take up before it's been calculated, but it's trivial to check how much RAM it takes up after it has been calculated.

This means that once the RAM ceiling is reached, the amount of RAM taken up is always going to be "one node" over budget.